refactor(kimi-k2): split NCCL MoE backend into moe_nccl.rs
The PPLX MoE backend already lived in runner/moe_pplx.rs while the NCCL
all-gather/reduce-scatter backend was mixed into worker/forward.rs. Extract
the NCCL MoE path into a sibling runner/moe_nccl.rs so the two backends sit
side by side and their only real difference -- cross-rank token movement --
is visible at a glance.

- moe_nccl.rs: the 4 NCCL MoE fns moved verbatim (decode / batch-prefill /
  prompt_len1-prefill), ordered decode-first to mirror moe_pplx.rs, with a
  doc table contrasting the two backends.
- runtime.rs: relocate the shared all_reduce_bf16_rows_in_place here; widen
  the NCCL collective helpers + kimi_marlin_block_size to pub(in crate::runner).
- worker.rs: re-export those helpers for moe_nccl; drop 19 now-unused imports.
- forward.rs / state.rs: repoint the 3 call sites to crate::runner::moe_nccl.
- moe_pplx.rs: fix a stale doc cross-reference to the moved NCCL path.

No behavior change: the moved fns are byte-identical to before (only the
decode entry point's visibility is widened). Verified with
cargo clippy -p pegainfer-kimi-k2 --features kimi-k2 -- -D warnings.
