+++ pegainfer-kimi-k2/src/runner.rs
@@ -3,6 +3,7 @@ mod bringup;
 mod config;
 mod executor;
 mod load_balancer;
+mod moe_nccl;
 mod moe_pplx;
 mod scheduler;
 mod worker;
