+++ pegainfer-kimi-k2/src/runner/worker/forward.rs
@@ -1,30 +1,5 @@
 use super::{runtime::*, *};
 
-pub(super) fn all_reduce_bf16_rows_in_place(
-    values: &mut CudaSlice<half::bf16>,
-    rows: usize,
-    row_len: usize,
-    comm: &Comm,
-) -> Result<()> {
-    ensure!(
-        values.len() >= rows * row_len,
-        "Kimi row-wise bf16 all-reduce len {} < rows {} * row_len {}",
-        values.len(),
-        rows,
-        row_len
-    );
-    for row in 0..rows {
-        let start = row * row_len;
-        let end = start + row_len;
-        let mut view = values.slice_mut(start..end);
-        comm.all_reduce_in_place(&mut view, &ReduceOp::Sum)
-            .map_err(|err| {
-                anyhow::anyhow!("Kimi row-wise bf16 all-reduce failed: status={:?}", err.0)
-            })?;
-    }
-    Ok(())
-}
-
 pub(super) fn forward_decode_batch_next_token_kernels(
     device_ctx: &DeviceContext,
     decode_aux_ctx: &DeviceContext,
@@ -106,7 +81,7 @@ pub(super) fn forward_decode_batch_next_token_kernels(
                         format!("Kimi MoE PPLX batch decode layer {}", layer.layer_idx)
                     })?;
                 } else {
-                    forward_moe_layer_decode_normed_into(
+                    crate::runner::moe_nccl::forward_moe_layer_decode_normed_into(
                         device_ctx,
                         decode_aux_ctx,
                         comm,
@@ -491,489 +466,6 @@ fn forward_dense_mlp_decode_normed_into(
         ctx,
         &scratch.mla.hidden,
         &scratch.mla.projected,
-        &mut scratch.mla.normed,
-    )?;
-    std::mem::swap(&mut scratch.mla.hidden, &mut scratch.mla.normed);
-    Ok(())
-}
-
-#[allow(clippy::too_many_arguments)]
-pub(super) fn forward_moe_layer_batch_into(
-    ctx: &DeviceContext,
-    comm: Option<&Comm>,
-    layer_idx: usize,
-    moe: &KimiMoeForwardCache,
-    post_attention_norm: &NormWeight<KIMI_K2_HIDDEN>,
-    expert_kernels: &KimiRankExpertMarlinWeights,
-    hidden: &mut GpuTensor<KIMI_K2_HIDDEN>,
-    normed: &mut GpuTensor<KIMI_K2_HIDDEN>,
-    next_hidden: &mut GpuTensor<KIMI_K2_HIDDEN>,
-) -> Result<()> {
-    let seq_len = hidden.seq_len;
-    typed_ops::rms_norm_into(
-        ctx,
-        hidden,
-        post_attention_norm,
-        KIMI_K2_RMS_NORM_EPS,
-        normed,
-    )?;
-    let mut shared_gate_up = HiddenStates::zeros(ctx, moe.shared_gate_up_proj.rows, seq_len)?;
-    typed_ops::gemm_dm_typed_to_hs(ctx, &moe.shared_gate_up_proj, normed, &mut shared_gate_up)?;
-    let mut shared_activated = HiddenStates::zeros(ctx, moe.shared_down_proj.cols, seq_len)?;
-    typed_ops::silu_mul_hs_fused_into(ctx, &shared_gate_up, &mut shared_activated)?;
-    let mut shared_out = GpuTensor::<KIMI_K2_HIDDEN>::zeros(ctx, seq_len)?;
-    typed_ops::gemm_dm_hs_to_typed(
-        ctx,
-        &moe.shared_down_proj,
-        &shared_activated,
-        &mut shared_out,
-    )?;
-    if let Some(comm) = comm {
-        comm.all_reduce_in_place(&mut shared_out.data, &ReduceOp::Sum)
-            .map_err(|err| {
-                anyhow::anyhow!("Kimi TP all-reduce bf16 hidden failed: status={:?}", err.0)
-            })?;
-    }
-
-    let mut router_logits = ctx.stream.alloc_zeros(seq_len * KIMI_K2_ROUTED_EXPERTS)?;
-    let mut router_scores = ctx.stream.alloc_zeros(seq_len * KIMI_K2_ROUTED_EXPERTS)?;
-    let mut router_choice_scores = ctx.stream.alloc_zeros(seq_len * KIMI_K2_ROUTED_EXPERTS)?;
-    let mut router_topk_weight = ctx.stream.alloc_zeros(seq_len * KIMI_K2_TOPK)?;
-    let mut router_topk_idx = ctx.stream.alloc_zeros(seq_len * KIMI_K2_TOPK)?;
-    {
-        let mut scratch = KimiRouterScratch {
-            logits: &mut router_logits,
-            scores: &mut router_scores,
-            choice_scores: &mut router_choice_scores,
-        };
-        let mut output = KimiRouterOutput {
-            topk_weight: &mut router_topk_weight,
-            topk_idx: &mut router_topk_idx,
-        };
-        kimi_router_noaux_tc_launch(
-            ctx,
-            KimiRouterConfig::kimi_k2(),
-            KimiRouterBatch {
-                batch_size: seq_len,
-                active_tokens: seq_len,
-                padded_tokens: seq_len,
-            },
-            normed,
-            &moe.router.gate_weight,
-            &moe.router.e_score_correction_bias,
-            &mut scratch,
-            &mut output,
-        )?;
-    }
-
-    let marlin_block_size = kimi_marlin_block_size(seq_len);
-    let mut route_workspace = KimiMarlinRouteWorkspace::new(ctx, seq_len, marlin_block_size)?;
-    let routing = kimi_moe_marlin_align_block_size(
-        ctx,
-        &mut route_workspace,
-        &router_topk_idx,
-        seq_len,
-        seq_len,
-        expert_kernels.local_expert_range.start,
-    )?;
-    let layer_weights = expert_kernels
-        .layers
-        .iter()
-        .find(|layer| layer.layer_idx == layer_idx)
-        .ok_or_else(|| {
-            anyhow::anyhow!("Kimi rank expert Marlin package missing layer {layer_idx}")
-        })?
-        .as_marlin_weights();
-
-    let mut marlin_workspace = KimiMarlinWna16Workspace::new(
-        ctx,
-        routing.max_m_blocks,
-        KIMI_K2_HIDDEN,
-        marlin_block_size,
-    )?;
-    let mut w13_out = GpuTensor::<MARLIN_W13_OUT_DIM>::zeros(ctx, routing.route_elems)?;
-    kimi_marlin_wna16_w13_gemm(
-        ctx,
-        &mut marlin_workspace,
-        &routing,
-        normed,
-        &layer_weights.w13,
-        &router_topk_weight,
-        &mut w13_out,
-    )?;
-    let mut activated = GpuTensor::<KIMI_K2_EXPERT_INTERMEDIATE>::zeros(ctx, routing.route_elems)?;
-    kimi_marlin_w13_swiglu(ctx, &w13_out, &mut activated)?;
-    let mut expert_output = GpuTensor::<KIMI_K2_HIDDEN>::zeros(ctx, routing.route_elems)?;
-    kimi_marlin_wna16_w2_gemm(
-        ctx,
-        &mut marlin_workspace,
-        &routing,
-        &activated,
-        &layer_weights.w2_down,
-        &router_topk_weight,
-        &mut expert_output,
-    )?;
-
-    let mut routed_out_f32 = ctx.stream.alloc_zeros(seq_len * KIMI_K2_HIDDEN)?;
-    kimi_marlin_sum_topk_rows_f32(ctx, &expert_output, seq_len, &mut routed_out_f32)?;
-    let nccl_comm = comm.ok_or_else(|| {
-        anyhow::anyhow!("NCCL MoE batch routed path requires TP comm (use PPLX for TP1)")
-    })?;
-    all_reduce_f32_in_place(&mut routed_out_f32, nccl_comm)?;
-    scale_f32_in_place(
-        ctx,
-        &mut routed_out_f32,
-        seq_len * KIMI_K2_HIDDEN,
-        KIMI_K2_ROUTER_SCALE,
-    )?;
-    pegainfer_kernels::typed_pipeline! {
-        ctx = ctx, eps = KIMI_K2_RMS_NORM_EPS;
-        add(hidden, &shared_out => next_hidden);
-    }
-    kimi_add_f32_bf16_to_bf16(ctx, &routed_out_f32, next_hidden, hidden)?;
-    Ok(())
-}
-
-pub(super) fn forward_moe_layer_prefill_scratch_into(
-    ctx: &DeviceContext,
-    comm: Option<&Comm>,
-    layer_idx: usize,
-    moe: &KimiMoeForwardCache,
-    post_attention_norm: &NormWeight<KIMI_K2_HIDDEN>,
-    expert_kernels: &KimiRankExpertMarlinWeights,
-    scratch: &mut KimiWorkerDecodeScratch,
-) -> Result<()> {
-    let seq_len = scratch.mla.hidden.seq_len;
-    typed_ops::rms_norm_into(
-        ctx,
-        &scratch.mla.hidden,
-        post_attention_norm,
-        KIMI_K2_RMS_NORM_EPS,
-        &mut scratch.mla.normed,
-    )?;
-    typed_ops::gemm_dm_typed_to_hs_per_token(
-        ctx,
-        &moe.shared_gate_up_proj,
-        &scratch.mla.normed,
-        &mut scratch.shared_expert.gate_up,
-    )?;
-    typed_ops::silu_mul_hs_fused_into(
-        ctx,
-        &scratch.shared_expert.gate_up,
-        &mut scratch.shared_expert.activated,
-    )?;
-    typed_ops::gemm_dm_hs_to_typed_per_token(
-        ctx,
-        &moe.shared_down_proj,
-        &scratch.shared_expert.activated,
-        &mut scratch.mla.projected,
-    )?;
-    if let Some(comm) = comm {
-        all_reduce_bf16_rows_in_place(
-            &mut scratch.mla.projected.data,
-            seq_len,
-            KIMI_K2_HIDDEN,
-            comm,
-        )?;
-    }
-
-    {
-        let mut router_scratch = KimiRouterScratch {
-            logits: &mut scratch.router.router_logits.data,
-            scores: &mut scratch.router.router_scores.data,
-            choice_scores: &mut scratch.router.router_choice_scores.data,
-        };
-        let mut router_output = KimiRouterOutput {
-            topk_weight: &mut scratch.router.router_topk_weight.data,
-            topk_idx: &mut scratch.router.router_topk_idx.data,
-        };
-        let router_batch = KimiRouterBatch {
-            batch_size: seq_len,
-            active_tokens: seq_len,
-            padded_tokens: seq_len,
-        };
-        kimi_router_noaux_tc_per_token_launch(
-            ctx,
-            KimiRouterConfig::kimi_k2(),
-            router_batch,
-            &scratch.mla.normed,
-            &moe.router.gate_weight,
-            &moe.router.e_score_correction_bias,
-            &mut router_scratch,
-            &mut router_output,
-        )?;
-    }
-
-    let routing = kimi_moe_marlin_align_block_size(
-        ctx,
-        &mut scratch.prompt_len1_moe.route_workspace,
-        &scratch.router.router_topk_idx.data,
-        seq_len,
-        seq_len,
-        expert_kernels.local_expert_range.start,
-    )?;
-    let layer_weights = expert_kernels
-        .layers
-        .iter()
-        .find(|layer| layer.layer_idx == layer_idx)
-        .ok_or_else(|| {
-            anyhow::anyhow!("Kimi rank expert Marlin package missing layer {layer_idx}")
-        })?
-        .as_marlin_weights();
-
-    scratch.marlin.w13_out.seq_len = routing.route_elems;
-    scratch.marlin.activated.seq_len = routing.route_elems;
-    scratch.marlin.expert_output.seq_len = routing.route_elems;
-    {
-        let mut active_w13 = scratch
-            .marlin
-            .w13_out
-            .data
-            .slice_mut(0..routing.route_elems * MARLIN_W13_OUT_DIM);
-        ctx.stream
-            .memset_zeros(&mut active_w13)
-            .with_context(|| format!("Kimi MoE layer {layer_idx} zero prompt w13 scratch"))?;
-    }
-    kimi_marlin_wna16_w13_gemm(
-        ctx,
-        &mut scratch.prompt_len1_moe.marlin_workspace,
-        &routing,
-        &scratch.mla.normed,
-        &layer_weights.w13,
-        &scratch.router.router_topk_weight.data,
-        &mut scratch.marlin.w13_out,
-    )?;
-    kimi_marlin_w13_swiglu(ctx, &scratch.marlin.w13_out, &mut scratch.marlin.activated)?;
-    {
-        let mut active_expert = scratch
-            .marlin
-            .expert_output
-            .data
-            .slice_mut(0..routing.route_elems * KIMI_K2_HIDDEN);
-        ctx.stream
-            .memset_zeros(&mut active_expert)
-            .with_context(|| format!("Kimi MoE layer {layer_idx} zero prompt expert scratch"))?;
-    }
-    kimi_marlin_wna16_w2_gemm(
-        ctx,
-        &mut scratch.prompt_len1_moe.marlin_workspace,
-        &routing,
-        &scratch.marlin.activated,
-        &layer_weights.w2_down,
-        &scratch.router.router_topk_weight.data,
-        &mut scratch.marlin.expert_output,
-    )?;
-
-    kimi_marlin_sum_topk_rows_f32(
-        ctx,
-        &scratch.marlin.expert_output,
-        seq_len,
-        &mut scratch.comm.routed_out_f32,
-    )?;
-    let nccl_comm = comm.ok_or_else(|| {
-        anyhow::anyhow!("NCCL MoE batch routed path requires TP comm (use PPLX for TP1)")
-    })?;
-    all_reduce_f32_rows_in_place(
-        &mut scratch.comm.routed_out_f32,
-        seq_len,
-        KIMI_K2_HIDDEN,
-        nccl_comm,
-    )?;
-    scale_f32_in_place(
-        ctx,
-        &mut scratch.comm.routed_out_f32,
-        seq_len * KIMI_K2_HIDDEN,
-        KIMI_K2_ROUTER_SCALE,
-    )?;
-    typed_ops::add_into(
-        ctx,
-        &scratch.mla.hidden,
-        &scratch.mla.projected,
-        &mut scratch.mla.normed,
-    )?;
-    kimi_add_f32_bf16_to_bf16(
-        ctx,
-        &scratch.comm.routed_out_f32,
-        &scratch.mla.normed,
-        &mut scratch.mla.hidden,
-    )?;
-    Ok(())
-}
-
-fn forward_moe_layer_decode_normed_into(
-    ctx: &DeviceContext,
-    aux_ctx: &DeviceContext,
-    comm: Option<&Comm>,
-    layer_idx: usize,
-    moe: &KimiMoeForwardCache,
-    expert_kernels: &KimiRankExpertMarlinWeights,
-    scratch: &mut KimiWorkerDecodeScratch,
-) -> Result<()> {
-    let norm_ready = ctx
-        .stream
-        .record_event(None)
-        .with_context(|| format!("Kimi MoE layer {layer_idx} record fused norm_ready"))?;
-    aux_ctx
-        .stream
-        .wait(&norm_ready)
-        .with_context(|| format!("Kimi MoE layer {layer_idx} aux wait fused norm_ready"))?;
-    forward_moe_layer_decode_normed_after_event_into(
-        ctx,
-        aux_ctx,
-        comm,
-        layer_idx,
-        moe,
-        expert_kernels,
-        scratch,
-    )
-}
-
-fn forward_moe_layer_decode_normed_after_event_into(
-    ctx: &DeviceContext,
-    aux_ctx: &DeviceContext,
-    comm: Option<&Comm>,
-    layer_idx: usize,
-    moe: &KimiMoeForwardCache,
-    expert_kernels: &KimiRankExpertMarlinWeights,
-    scratch: &mut KimiWorkerDecodeScratch,
-) -> Result<()> {
-    let seq_len = scratch.mla.hidden.seq_len;
-    typed_ops::gemm_dm_typed_to_hs_graphsafe(
-        ctx,
-        &moe.shared_gate_up_proj,
-        &scratch.mla.normed,
-        &mut scratch.shared_expert.gate_up,
-    )?;
-    typed_ops::silu_mul_hs_fused_into(
-        ctx,
-        &scratch.shared_expert.gate_up,
-        &mut scratch.shared_expert.activated,
-    )?;
-    typed_ops::gemm_dm_hs_to_typed_graphsafe(
-        ctx,
-        &moe.shared_down_proj,
-        &scratch.shared_expert.activated,
-        &mut scratch.mla.projected,
-    )?;
-    maybe_all_reduce_hidden_via_f32_in_place(
-        ctx,
-        &mut scratch.mla.projected,
-        &mut scratch.comm.hidden_allreduce_f32,
-        comm,
-    )?;
-
-    // Router + routed experts (aux stream)
-    {
-        let mut router_scratch = KimiRouterScratch {
-            logits: &mut scratch.router.router_logits.data,
-            scores: &mut scratch.router.router_scores.data,
-            choice_scores: &mut scratch.router.router_choice_scores.data,
-        };
-        let mut router_output = KimiRouterOutput {
-            topk_weight: &mut scratch.router.router_topk_weight.data,
-            topk_idx: &mut scratch.router.router_topk_idx.data,
-        };
-        kimi_router_noaux_tc_launch(
-            aux_ctx,
-            KimiRouterConfig::kimi_k2(),
-            KimiRouterBatch {
-                batch_size: seq_len,
-                active_tokens: seq_len,
-                padded_tokens: seq_len,
-            },
-            &scratch.mla.normed,
-            &moe.router.gate_weight,
-            &moe.router.e_score_correction_bias,
-            &mut router_scratch,
-            &mut router_output,
-        )?;
-    }
-    let routing = kimi_moe_marlin_align_block_size(
-        aux_ctx,
-        &mut scratch.marlin_route_workspace,
-        &scratch.router.router_topk_idx.data,
-        seq_len,
-        seq_len,
-        expert_kernels.local_expert_range.start,
-    )?;
-    let layer_weights = expert_kernels
-        .layers
-        .iter()
-        .find(|layer| layer.layer_idx == layer_idx)
-        .ok_or_else(|| {
-            anyhow::anyhow!("Kimi rank expert Marlin package missing layer {layer_idx}")
-        })?
-        .as_marlin_weights();
-
-    aux_ctx
-        .stream
-        .memset_zeros(&mut scratch.marlin.w13_out.data)?;
-    kimi_marlin_wna16_w13_gemm(
-        aux_ctx,
-        &mut scratch.marlin_workspace,
-        &routing,
-        &scratch.mla.normed,
-        &layer_weights.w13,
-        &scratch.router.router_topk_weight.data,
-        &mut scratch.marlin.w13_out,
-    )?;
-    kimi_marlin_w13_swiglu(
-        aux_ctx,
-        &scratch.marlin.w13_out,
-        &mut scratch.marlin.activated,
-    )?;
-    aux_ctx
-        .stream
-        .memset_zeros(&mut scratch.marlin.expert_output.data)?;
-    kimi_marlin_wna16_w2_gemm(
-        aux_ctx,
-        &mut scratch.marlin_workspace,
-        &routing,
-        &scratch.marlin.activated,
-        &layer_weights.w2_down,
-        &scratch.router.router_topk_weight.data,
-        &mut scratch.marlin.expert_output,
-    )?;
-    kimi_marlin_sum_topk_rows_f32(
-        aux_ctx,
-        &scratch.marlin.expert_output,
-        seq_len,
-        &mut scratch.comm.routed_out_f32,
-    )?;
-    repeat_f32_for_reduce_scatter_into(
-        aux_ctx,
-        &scratch.comm.routed_out_f32,
-        &mut scratch.comm.routed_reduce_scatter_send_f32,
-        seq_len * KIMI_K2_HIDDEN,
-        KIMI_K2_EP_WORLD,
-    )?;
-
-    let routed_local_done = aux_ctx
-        .stream
-        .record_event(None)
-        .with_context(|| format!("Kimi MoE layer {layer_idx} record routed_local_done"))?;
-    ctx.stream
-        .wait(&routed_local_done)
-        .with_context(|| format!("Kimi MoE layer {layer_idx} main wait routed_local_done"))?;
-    let nccl_comm = comm.ok_or_else(|| {
-        anyhow::anyhow!("NCCL MoE routed path requires TP comm (use PPLX for TP1)")
-    })?;
-    reduce_scatter_f32_hidden_into(
-        &scratch.comm.routed_reduce_scatter_send_f32,
-        seq_len * KIMI_K2_EP_WORLD,
-        KIMI_K2_HIDDEN,
-        &mut scratch.comm.routed_out_f32,
-        seq_len,
-        KIMI_K2_EP_WORLD,
-        nccl_comm,
-    )?;
-
-    kimi_residual_add_scaled_f32(
-        ctx,
-        &scratch.mla.hidden,
-        &scratch.mla.projected,
-        &scratch.comm.routed_out_f32,
-        KIMI_K2_ROUTER_SCALE,
         &mut scratch.mla.normed,
     )?;
     std::mem::swap(&mut scratch.mla.hidden, &mut scratch.mla.normed);
